requirements.txt
flask
pystray
pillow
requests
pyperclip
